Move `retry_with_navigation` from mcp/utils.py to browser.py
Consolidates the function with the other navigation utilities it depends on
(zen_navigate_with_retry), removing the awkward lazy import it needed to
avoid a circular dependency. Deletes `mcp/utils.py`.
