+++ getgather/mcp/tokopedia.py
@@ -6,13 +6,12 @@ from urllib.parse import quote, urlparse
 import zendriver as zd
 from loguru import logger
 
-from getgather.browser import page_query_selector, zen_navigate_with_retry
+from getgather.browser import page_query_selector, retry_with_navigation, zen_navigate_with_retry
 from getgather.mcp.dpage import (
     remote_zen_dpage_mcp_tool,
     remote_zen_dpage_with_action,
 )
 from getgather.mcp.registry import GatherMCP
-from getgather.mcp.utils import retry_with_navigation
 from getgather.zen_actions import parse_response_json
 
 tokopedia_mcp = GatherMCP(brand_id="tokopedia", name="Tokopedia MCP")
